Updated ReservNFT test to cover royalty logic
Setup: I updated the setUp function to include the two royalty parameters. This makes sure that any NFT minted from that drop does have the royalty info.

Testing: I modified the ‘test_createReservNFT’ function for testing that the contract properly associates the royalty info with the NFTs and can be pulled and verified.

After the NFT is created, the test pulls the royalty details for that NFT using the royaltyInfo function (ERC2981). I call that function with the token ID of the NFT and a 1000 Wei example sale price. I then include two assertions to check that the fetched royalty info matches what we setup.
- Check that the royalty receiver address that’s returned matches the test address.
- Check that the royalty amount returned for that sale price is right. (10 Wei in this case)